refactor: move hyperopt-auto
